pyproject.toml
[project]
name = "fastapilms"
version = "0.1.0"
description = ""
authors = [
    {name = "AntonSkyrda",email = "mr.skyrda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standart] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "invoke (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
